// src/components/proTable/table.types.ts
export type ProTablePageParams = {
  currentPage?: number
  pageSize?: number
  total?: number
  scrollId?: string
  hasMore?: boolean
}